Monsters now roam again

+++ src/mortal.hpp
@@ -2,14 +2,14 @@
 
 class Mortal {
 public:
-	float MaxHp, hp, def;
+	int MaxHp, hp, def;
 	const std::string corpseName;
 
-	Mortal(float MaxHp, float def, std::string corpseName);
+	Mortal(int MaxHp, int def, std::string corpseName);
 
 	bool isDead() { return hp <= 0; }
 
-	float takeDamage(std::shared_ptr<Entity> owner, float dmg);
+	int takeDamage(std::shared_ptr<Entity> owner, int dmg);
 
 	void die(std::shared_ptr<Entity> owner);
 };
